pyproject.toml
[project]
name = "fitonduty-dashboard-march"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.4",
    "flask-login>=0.6.3",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.2",
    "ruff>=0.12.12",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = ["E501", "B008"]
